Add missing newlines
On *nix systems it's a common convention to end a file with an empty
newline. It also helps if the newline is present if user wants to
simply cat a file in the command line.

I used the following one-liner to automatically add newlines in all
files that don't have it:

    $ find . -path ./.git -prune -o -type f -a ! -name .keep -exec sh -c '[[ $(tail -c1 "$1" | wc -l) -gt 0 ]] || echo >> "$1"' sh {} \;